requirements.txt
# Polymarket CLOB client
py-clob-client>=0.5.0

# Real-time data
websockets>=12.0
aiohttp>=3.9.0

# Data analysis
numpy>=1.26.0
pandas>=2.1.0

# Technical indicators
ta>=0.11.0

# Configuration
python-dotenv>=1.0.0

# HTTP
requests>=2.31.0

# Logging
structlog>=24.1.0

# Web3 (for Polymarket/Polygon)
web3>=6.15.0
eth-account>=0.11.0

# Database (Heroku Postgres)
psycopg2-binary>=2.9.9

# Web dashboard
flask>=3.0.0
gunicorn>=21.2.0
